requirements.txt
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.22.1
gunicorn==20.1.0
gevent>=22.10.0
phonenumbers==8.12.49
psycopg2-binary==2.9.2
pyjwt~=2.4.0
factory-boy==3.2.0
Faker==30.0.0
whitenoise==5.3.0
django-model-utils==4.3.1
django-filter==23.2
django-phonenumber-field==7.1.0